Created BaseDaggerComponent
This component will be implemented by all dagger components. It already
contains the abstract function that dagger will use for injection, as
well as a factory for creating the component.

The init function is used to inject the target, calling pre/post inject
if available and registering a dagger injector, if available, with
DaggerInjection, to be used to inject subcomponents.

When this component is created, init must be called to inject the target
and enable injection for all subcomponents.