Created BaseDaggerSubcomponent
A dagger subcomponent that will be used in conjunction with the
BaseDaggerComponent. The inject function is already included for dagger
to inject the target. The Factory class must be implemented to provide
the key, which will be the class of the target to inject.

The FactoryMapping must also be extended and added as a module to the
component, to provide the mapping required for DaggerInjector to work.