Created JSON web tokens
The JSONWebToken contains only the header and payload and their Base64
encoded Strings. The key id is automatically set in the header. The
token can be converted into its String form. Commonly used key
algorithms, headers and claims are also provided.

The JSONWebSignature also includes a signature with a signing method
that must be implemented by subclasses.

The HMACJSONWebSignature signs the token using HMAC SHA256.