[FEATURE] API Module (#4)
Resources represents an entity that can be stored and passed around. EmbeddableResources are resources that can also be embedded, its embedded version containing only the necessary information from the original resource.

ResourceConverters are converters that converts a source (such as a JSONObject) to a resource, and vice versa. Validators are used to a resource, before carrying on a request.

Endpoints responds to requests made. CallableEndpoint will perform only one task, while ResourceEndpoint will answer to requests that involves a Resource.
Response

EndpointResponses will be returned on any requests made to the endpoints. They can either be successful or unsuccessful. Endpoint failures can be handled by the EndpointFailureHandler.